fix: prod 제외 설정 적용
