+++ api/src/main/kotlin/com/retoday/api/global/config/SecurityConfiguration.kt
@@ -9,7 +9,6 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
@@ -45,23 +44,6 @@ class SecurityConfiguration {
             }
             addFilterAt(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
             build()
-        }
-
-    @Bean
-    fun webSecurityCustomizer(): WebSecurityCustomizer =
-        WebSecurityCustomizer { web ->
-            web
-                .ignoring()
-                .requestMatchers(
-                    "/v3/api-docs",
-                    "/v3/api-docs/**",
-                    "/swagger-ui/**",
-                    "/swagger-ui.html",
-                    "/swagger-resources/**",
-                    "/api-docs/**",
-                    "/favicon.ico",
-                    "/error"
-                )
         }
 
     @Bean
