+++ api/src/main/resources/application.yaml
@@ -10,6 +10,13 @@ spring:
   web:
     resources:
       add-mappings: true
+springdoc:
+  swagger-ui:
+    url: /docs/api.yaml
+    path: /swagger-ui.html
+    disable-swagger-default-url: true
+  api-docs:
+    enabled: true
 server:
   servlet:
     encoding:
@@ -47,16 +54,12 @@ spring:
   config:
     activate:
       on-profile: prod
+  web:
+    resources:
+      add-mappings: false
+
 springdoc:
-  default-consumes-media-type: application/json
-  default-produces-media-type: application/json
   swagger-ui:
-    path: /swagger-ui.html
-    disable-swagger-default-url: true
-    display-request-duration: true
-    groups-order: desc
-    operations-sorter: method
+    enabled: false
   api-docs:
-    path: /v3/api-docs
-    enabled: true
-  show-actuator: false
+    enabled: false
